[ADD] editable-text, project-menu
